+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/citizens/SKCitizensShopObjectType.java
@@ -6,12 +6,14 @@ import java.util.UUID;
 import org.bukkit.Location;
 import org.bukkit.block.BlockFace;
 import org.bukkit.entity.Entity;
+import org.bukkit.entity.Player;
 
 import com.nisovin.shopkeepers.api.shopkeeper.ShopCreationData;
 import com.nisovin.shopkeepers.api.shopobjects.citizens.CitizensShopObjectType;
 import com.nisovin.shopkeepers.lang.Messages;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
 import com.nisovin.shopkeepers.shopobjects.entity.AbstractEntityShopObjectType;
+import com.nisovin.shopkeepers.util.bukkit.TextUtils;
 
 public final class SKCitizensShopObjectType extends AbstractEntityShopObjectType<SKCitizensShopObject> implements CitizensShopObjectType<SKCitizensShopObject> {
 
@@ -61,9 +63,14 @@ public final class SKCitizensShopObjectType extends AbstractEntityShopObjectType
 	}
 
 	@Override
-	public boolean isValidSpawnLocation(Location spawnLocation, BlockFace targetedBlockFace) {
+	public boolean validateSpawnLocation(Player creator, Location spawnLocation, BlockFace targetedBlockFace) {
 		// A reduced set of checks, compared to the default:
-		if (spawnLocation == null || spawnLocation.getWorld() == null) return false;
+		if (spawnLocation == null || spawnLocation.getWorld() == null) {
+			if (creator != null) {
+				TextUtils.sendMessage(creator, Messages.missingSpawnLocation);
+			}
+			return false;
+		}
 		return true;
 	}